Move common methods to framework.go
